feat(topico): implementing the topic creation method.

+++ src/main/java/com/eranitor/tcc/entity/Materia.java
@@ -16,7 +16,7 @@ public class Materia {
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
     @Column (name = "idmateria")
-    private Integer idMateria;
+    private Long idMateria;
 
     @ManyToOne
     @JoinColumn (name = "usuario_idusuario", nullable = false)
