+++ src/main/java/com/eranitor/tcc/repository/MateriaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface MateriaRepository extends JpaRepository<Materia, Integer> {
+public interface MateriaRepository extends JpaRepository<Materia, Long> {
 
     Optional<Materia> findByNomeIgnoreCaseAndUsuario_IdUsuario(
             String nome,
